requirements.txt
pandas~=2.1.4
numpy~=1.23.4
torch~=2.4.0
torchvision~=0.19.0
opencv-python~=4.10.0.84
transformers~=4.44.2
Pillow~=9.5.0
PyYAML~=6.0